# Implement lazy SIR loading.