--- a/ykpack/Cargo.toml
+++ b/ykpack/Cargo.toml
@@ -9,4 +9,5 @@ license = "Apache-2.0 OR MIT"
 bincode = "1.3.1"
 bitflags = "1.2"
 fallible-iterator = "0.2.0"
+fxhash = "0.2.1"
 serde = { version = "1.0.115", features = ["derive"] }