@ltratt 2020-12-01
Those performance numbers suggest that we're now running parallel code more efficiently (as in: not block as often) because the `user` portion is more-or-less unchanged. Does that sound plausible?