@ptersilie 2020-12-01
After reading up on stacked borrows and playing around with MIRI I have addressed some of the comments. I've also refactored some of the code. So I think this is ready for re-review.