@vext01 2020-12-01
> Those performance numbers suggest that we're now running parallel code more efficiently.

I think so, because before all of the test threads had to block for the long-running `lazy_static` to load the SIR. Now the runtime of that `lazy_static` is very short, as we only load a handful of small headers in.